# mq/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Custom Message Queue Broker with topic-based pub/sub, consumer groups, and at-least-once delivery.",
        "title": "MQ Broker API",
        "contact": {
            "name": "GPU Telemetry Pipeline"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
...
        "/admin/topics/{topic}/consumers": {
            "get": {
                "description": "Returns all consumers for a topic",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get topic consumers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Topic name",
                        "name": "topic",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.ConsumersResponse"
                        }
                    }
                }
            }
        },
        "/admin/topics/{topic}/dlq": {
            "delete": {
                "description": "Deletes all messages from the Dead Letter Queue for a topic",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "DLQ"
                ],
                "summary": "Purge DLQ messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Original topic name (without -dlq suffix)",
                        "name": "topic",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.PurgeResponse"
                        }
                    }
                }
            }
        },
        "/admin/topics/{topic}/dlq/messages": {
            "get": {
                "description": "Returns messages in the Dead Letter Queue for a topic with failure details",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "DLQ"
                ],
                "summary": "Get DLQ messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Original topic name (without -dlq suffix)",
                        "name": "topic",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum messages to return (default 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.DLQMessagesResponse"
                        }
                    }
                }
            }
        },
        "/admin/topics/{topic}/dlq/replay": {
            "post": {
                "description": "Replays messages from the Dead Letter Queue back to the original topic",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "DLQ"
                ],
                "summary": "Replay DLQ messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Original topic name (without -dlq suffix)",
                        "name": "topic",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Optional: specific message IDs to replay",
                        "name": "replay",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.ReplayRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.ReplayResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/topics/{topic}/dlq/stats": {
            "get": {
                "description": "Returns Dead Letter Queue statistics for a topic",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "DLQ"
                ],
                "summary": "Get DLQ statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Original topic name (without -dlq suffix)",
                        "name": "topic",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.DLQStats"
                        }
                    }
                }
            }
        },
        "/admin/topics/{topic}/messages": {
            "get": {
                "description": "Returns messages from a topic for administrative viewing",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get messages from topic (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Topic name",
                        "name": "topic",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum messages to return (default 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.MessagesResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes all messages from a topic",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Purge all messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Topic name",
                        "name": "topic",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.PurgeResponse"
                        }
                    }
                }
            }
        },
        "/admin/topics/{topic}/messages/{messageID}": {
            "delete": {
                "description": "Deletes a specific message from a topic",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Delete a message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Topic name",
                        "name": "topic",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "messageID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.DeleteResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/topics/{topic}/stats": {
            "get": {
                "description": "Returns statistics for a specific topic",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get topic statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Topic name",
                        "name": "topic",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.TopicStats"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/topics/{topic}/ack": {
            "post": {
                "description": "Acknowledges receipt of messages by a consumer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Acknowledge messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Topic name",
                        "name": "topic",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Acknowledgment details",
                        "name": "ack",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.AckRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.AckResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/topics/{topic}/subscribe": {
            "post": {
                "description": "Subscribes a consumer to a topic within a consumer group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Consumers"
                ],
                "summary": "Subscribe to a topic",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Topic name",
                        "name": "topic",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Subscription details",
                        "name": "subscription",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.SubscribeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.SubscribeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_gpu-telemetry-pipeline_mq_pkg_models.ConsumersResponse": {
            "description": "Consumers for a topic",
            "type": "object",
            "properties": {
                "consumers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.ConsumerInfo"
                    }
                },
                "topic": {
                    "type": "string",
                    "example": "gpu-telemetry"
                }
            }
        },
...
        "github_com_gpu-telemetry-pipeline_mq_pkg_models.DLQMessagesResponse": {
            "description": "DLQ messages response",
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 10
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_gpu-telemetry-pipeline_mq_pkg_models.DLQMessage"
                    }
                },
                "topic": {
                    "type": "string",
                    "example": "gpu-telemetry-dlq"
                }
            }
        },
...
        "github_com_gpu-telemetry-pipeline_mq_pkg_models.ErrorResponse": {
            "description": "Error response",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string",
                    "example": "topic not found"
                }
            }
        },
...
        "github_com_gpu-telemetry-pipeline_mq_pkg_models.PurgeResponse": {
            "description": "Purge response",
            "type": "object",
            "properties": {
                "messages_deleted": {
                    "type": "integer",
                    "example": 100
                },
                "status": {
                    "type": "string",
                    "example": "purged"
                }
            }
        },
        "github_com_gpu-telemetry-pipeline_mq_pkg_models.ReplayRequest": {
            "description": "Request to replay DLQ messages",
            "type": "object",
            "properties": {
                "force": {
                    "description": "If true, also replay dead messages",
                    "type": "boolean"
                },
                "message_ids": {
                    "description": "If empty, replay all pending DLQ messages",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_gpu-telemetry-pipeline_mq_pkg_models.ReplayResponse": {
            "description": "Response after replaying DLQ messages",
            "type": "object",
            "properties": {
                "failed": {
                    "type": "integer",
                    "example": 0
                },
                "replayed": {
                    "type": "integer",
                    "example": 5
                },
                "status": {
                    "type": "string",
                    "example": "replayed"
                }
            }
        },
...
    }
}